More css and styling updates to agent config
